UML-10 cleanup
