--- src/app/types.ts
@@ -1,6 +1,0 @@
-import { AppLocales } from '../locales/types';
-
-export interface AppType {
-    getLocale(): AppLocales
-    setLocale(locale: AppLocales): void
-}
