--- src/components/uml-editor/Widget/types.ts
@@ -1,5 +1,0 @@
-import { WidgetStore } from './WidgetStore';
-
-export type WidgetProps = {
-    widget: WidgetStore
-};
