+++ src/components/uml-editor/items/types.ts
@@ -2,13 +2,13 @@ import { I18n } from '@lingui/core';
 
 export type ItemType = 'class' | 'interface';
 
+export type ItemProps = {
+    i18n: I18n
+};
+
 export type ItemStruct<Type extends ItemType = ItemType> = {
     type: Type
     name: string
-};
-
-export type ItemProps = {
-    i18n: I18n
 };
 
 export interface ItemsFactoryI {
