+++ src/locales/I18nAppProvider.tsx
@@ -1,8 +1,12 @@
 import React, { useEffect } from 'react';
 import { I18nProvider } from '@lingui/react';
 import { i18n } from '@lingui/core';
-import { dynamicActivate } from './i18n';
-import { I18nAppProviderProps } from './types';
+import { dynamicActivate, AppLocales } from './i18n';
+
+type I18nAppProviderProps = {
+    children: React.ReactNode
+    locale: AppLocales
+};
 
 const I18nAppProvider = (props: I18nAppProviderProps) => {
     const { children, locale } = props;
