--- src/router/config/types.ts
@@ -1,11 +1,0 @@
-type RouteType = {
-    id: string
-    path: string
-    name: string,
-    loadComponent(): Promise<{readonly default: () => JSX.Element}>,
-    routes?: RouteType[]
-};
-
-export type {
-    RouteType,
-};
