--- src/router/route-provider/types.ts
@@ -1,7 +1,0 @@
-import { RouteType } from '../config/types';
-
-export type RouteListProps = {
-    list: RouteType[]
-};
-
-export type LoadFn = () => Promise<{readonly default: () => JSX.Element}>;
