--- src/utils/Broker/Broker.ts
@@ -1,10 +1,0 @@
-class Broker {
-    private e: number;
-
-    constructor() {
-        this.e = 1;
-
-    }
-}
-
-export default Broker;
